refactor: rename oauth to oauth-lc
