+++ api/ticket/api.js
@@ -2,7 +2,7 @@ const AV = require('leanengine')
 const { Router } = require('express')
 const { check, query } = require('express-validator')
 const { captureException } = require('../errorHandler')
-const { checkPermission } = require('../oauth')
+const { checkPermission } = require('../oauth-lc')
 const { requireAuth, catchError, parseSearchingQ } = require('../middleware')
 const { getVacationerIds, getFormValuesDifference } = require('./utils')
 const { isObjectExists } = require('../utils/object')
